7 methodology to analysis songs
Using 7 machine learning methodology to analysis songs to predict whether it will be popular when it comes to a new song.
The result is not ideal. Because the songs is so various and complicated. It's hard to figure out the rule of art.  Besides, we consider that the songs we collected are all from one year in one specified field - North America, which also increase the difficulty to classify. Therefore, the future work can be comparing the songs from different ages or different field. Maybe that works.